+++ packages/core/inline-message/examples/03-type-connectivity.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import InlineMessage from '../src';
 
 const MessageContent = (